`window\settings`: rewrite internal structure
