+++ susapad/window/settings/component/__init__.py
@@ -1,0 +1,16 @@
+from .buttons  import CloseButton as Close
+from .sliders  import ActuationPointGroup as ActuationPoint
+from .sliders  import SensiblitySlidersGroup as Sensiblity
+from .togglers import TriggerButton as Trigger
+from .togglers import RapidTriggerButton as RapidTrigger
+from .togglers import ContinuousRapidTriggerButton as ContinuousRapidTrigger
+
+
+__all__ = [
+    "Close",
+    "Trigger",
+    "RapidTrigger",
+    "ContinuousRapidTrigger",
+    "ActuationPoint",
+    "Sensiblity"
+]
